Issue #12206: Update page tests as per the changes
